test_runner: reimplement `assert.ok` to allow stack parsing
