+++ lib/internal/test_runner/test.js
@@ -64,6 +64,7 @@ const { setTimeout } = require('timers');
 const { TIMEOUT_MAX } = require('internal/timers');
 const { fileURLToPath } = require('internal/url');
 const { availableParallelism } = require('os');
+const { innerOk } = require('internal/assert/utils');
 const { bigint: hrtime } = process.hrtime;
 const kCallbackAndPromisePresent = 'callbackAndPromisePresent';
 const kCancelledByParent = 'cancelledByParent';
@@ -113,7 +114,6 @@ function lazyAssertObject(harness) {
       'notDeepStrictEqual',
       'notEqual',
       'notStrictEqual',
-      'ok',
       'rejects',
       'strictEqual',
       'throws',
@@ -263,6 +263,16 @@ class TestContext {
           return ReflectApply(method, this, args);
         };
       });
+
+      // This is a hack. It allows the innerOk function to collect the stacktrace from the correct starting point.
+      function ok(...args) {
+        if (plan !== null) {
+          plan.actual++;
+        }
+        innerOk(ok, args.length, ...args);
+      }
+
+      assert.ok = ok;
     }
     return this.#assert;
   }
