+++ test/parallel/test-runner-assert.js
@@ -24,3 +24,7 @@ test('only methods from node:assert are on t.assert', (t) => {
     'throws',
   ]);
 });
+
+test('t.assert.ok correctly parses the stacktrace', (t) => {
+  t.assert.throws(() => t.assert.ok(1 === 2), /t\.assert\.ok\(1 === 2\)/);
+});
